lift tabs logic into context
